sai: Delay SAI DMA TX disable until FIFO is empty
This resolves a bug that otherwise would cause the last frame to be
written to the FIFO but the FIFO and transmitter are disable before the
frame is sent.  The frame will be sent on the next enable of the
transmitter.

Instead, disable the transmitter after the FIFO has consumed the last
frame.